Add function to play a round of the game, fix bug
Yes bug singular, apparently the if ... else statement is itself another
scope different than the scope in which it was created. Long story short,
when trying to return playerChoice in the getPlayerChoice function I was
getting an error due to scope.

About the main function of the whole project, playRound, I am sure that
it can be done in a more efficient way, but alas, I am but humble
student. The more efficient and complex ways of javascript and coding in
general will come to me given enough time.