Add function playGame and 'fix' function playRound
Actually read the documentation for the project and fixed some things
that WEREN'T WRONG, but also weren't what the project asked for.

If a had one nitpick it would be that playRound *does* console.log the
winner for the current round, when the documentation says to return it.
I just couldn't think of a better way pass the to playGame function the
data about who won the round (and checking for the string literal seemed
too cheap to me).